+++ l10n_it_financial_statement_eu/static/src/xml/report.xml
@@ -7,18 +7,6 @@
         owl="1"
     >
         <xpath expr="//button" position="after">
-<!--            <button-->
-<!--                t-on-click="on_click_xlsx"-->
-<!--                type="button"-->
-<!--                class="btn btn-secondary"-->
-<!--                title="Export XLSX"-->
-<!--            >Export XLSX</button>-->
-<!--            <button-->
-<!--                t-on-click="on_click_xlsx"-->
-<!--                type="button"-->
-<!--                class="btn btn-secondary"-->
-<!--                title="Export XLSX"-->
-<!--            >Export XBRL</button>-->
             <button
                 t-if="isFinancialStatementEU"
                 type="button"
